CLI-123: Extract config file processing common logic
