+++ lib/EnvFileProcessor.js
@@ -17,6 +17,7 @@ const async = require('async');
 
 const { AllowedPermissionsOperations, AppOptionsName, BackendCollectionPermission, CollectionHook, DomainTypes,
   EntityType, EnvOptionsName, LogLevel, OperationType } = require('./Constants');
+const FileProcessorHelper = require('./FileProcessorHelper');
 const { Endpoints, getCustomNotFoundError, getObjectByOmitting, isEmpty, isNullOrUndefined } = require('./Utils');
 
 class EnvFileProcessor {
@@ -277,8 +278,8 @@ class EnvFileProcessor {
     });
   }
 
-  _groupCollectionsPerOperationType(originalCollections, collections) {
-    const groupedEntities = this._groupEntitiesPerOperationType(originalCollections, collections);
+  static _groupCollectionsPerOperationType(originalCollections, collections) {
+    const groupedEntities = FileProcessorHelper.groupEntitiesPerOperationType(originalCollections, collections);
     if (!isEmpty(groupedEntities[OperationType.DELETE])) {
       const collToDel = groupedEntities[OperationType.DELETE].filter(x => x !== 'user' && x !== '_blob');
       groupedEntities[OperationType.DELETE] = collToDel;
@@ -418,7 +419,7 @@ class EnvFileProcessor {
             return next(err);
           }
 
-          groupedEntities = this._groupCollectionsPerOperationType(originalCollections, collections);
+          groupedEntities = EnvFileProcessor._groupCollectionsPerOperationType(originalCollections, collections);
           next();
         });
       },
@@ -522,7 +523,7 @@ class EnvFileProcessor {
           }
 
           existingEndpoints = originalEndpoints;
-          groupedEntities = this._groupEntitiesPerOperationType(originalEndpoints, endpoints);
+          groupedEntities = FileProcessorHelper.groupEntitiesPerOperationType(originalEndpoints, endpoints);
 
           next();
         });
@@ -593,60 +594,6 @@ class EnvFileProcessor {
     ], done);
   }
 
-  /**
-   * Groups modifiedEntities per operation type: create, update, delete.
-   * @param {Array} originalEntities A list of strings or a list of objects with a 'name' property or other property of choice.
-   * @param {Object} modifiedEntities Keys on first level are the entity's name. Ex.: { "someName": {...}, "anyName": {...} }
-   * @param {String} [prop] Property to group by. Defaults to 'name'.
-   * @returns {Object}
-   * @private
-   */
-  // eslint-disable-next-line
-  _groupEntitiesPerOperationType(originalEntities, modifiedEntities, prop = 'name') {
-    const entitiesToDelete = [];
-    const entitiesToCreate = {};
-    const entitiesToUpdate = [];
-
-    const entityNamesToModify = Object.keys(modifiedEntities);
-
-    originalEntities.forEach((originalEntity) => {
-      let originalName;
-      if (originalEntity && isNullOrUndefined(originalEntity[prop])) {
-        originalName = originalEntity;
-      } else {
-        originalName = originalEntity[prop];
-      }
-
-      if (!entityNamesToModify.includes(originalName)) {
-        entitiesToDelete.push(originalName);
-      } else {
-        entitiesToUpdate.push({
-          [originalName]: modifiedEntities[originalName]
-        });
-      }
-    });
-
-    entityNamesToModify.forEach((entityName) => {
-      const entityAlreadyExists = originalEntities.find((x) => {
-        if (!isNullOrUndefined(x[prop])) {
-          return x[prop] === entityName;
-        }
-
-        return x === entityName;
-      });
-
-      if (!entityAlreadyExists) {
-        entitiesToCreate[entityName] = modifiedEntities[entityName];
-      }
-    });
-
-    return {
-      [OperationType.CREATE]: entitiesToCreate,
-      [OperationType.UPDATE]: entitiesToUpdate,
-      [OperationType.DELETE]: entitiesToDelete
-    };
-  }
-
   _modifyCommonCode(commonCode = {}, env, done) {
     const envId = env.id;
     let groupedEntities;
@@ -658,7 +605,7 @@ class EnvFileProcessor {
             return next(err);
           }
 
-          groupedEntities = this._groupEntitiesPerOperationType(originalCommonModules, commonCode);
+          groupedEntities = FileProcessorHelper.groupEntitiesPerOperationType(originalCommonModules, commonCode);
           next();
         });
       },
